Features

+++ MediaStorage.Api/Controllers/FilesController.cs
@@ -3,10 +3,12 @@ using MediaStorage.Api.Infrastructure;
 using MediaStorage.Api.Services;
 using MediaStorage.Api.Domain;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Authorization;
 
 namespace MediaStorage.Api.Controllers;
 
 [ApiController]
+[Authorize]
 [Route("api/files")]
 public class FilesController(MediaStorageDbContext dbContext, BlobStorageService blobStorageService) : ControllerBase
 {
@@ -23,6 +25,12 @@ public class FilesController(MediaStorageDbContext dbContext, BlobStorageService
             return BadRequest("Category is required.");
         }
 
+        var role = User.FindFirst("role")?.Value;
+        if (category == "lectures" && role != "Teacher")
+        {
+            return Forbid();
+        }
+
         var fileId = Guid.NewGuid();
 
         var extension = Path.GetExtension(file.FileName);
@@ -33,10 +41,12 @@ public class FilesController(MediaStorageDbContext dbContext, BlobStorageService
 
         await blobStorageService.UploadAsync(stream, blobName, file.ContentType, cancellationToken);
 
+        var userId = Guid.Parse(User.FindFirst("uid")!.Value);
+
         var storedFile = new StoredFile
         {
             Id = fileId,
-            OwnerId = Guid.Empty, // change with JWT user id
+            OwnerId = userId,
             FileName = file.FileName,
             BlobName = blobName,
             ContentType = file.ContentType,
@@ -74,6 +84,12 @@ public class FilesController(MediaStorageDbContext dbContext, BlobStorageService
             return NotFound();
         }
 
+        var userId = Guid.Parse(User.FindFirst("uid")!.Value);
+        if (file.OwnerId != userId)
+        {
+            return Forbid();
+        }
+
         var response = new FileResponse
         {
             FileId = file.Id,
@@ -96,6 +112,13 @@ public class FilesController(MediaStorageDbContext dbContext, BlobStorageService
         if (file is null)
         {
             return NotFound();
+        }
+
+        var userId = Guid.Parse(User.FindFirst("uid")!.Value);
+
+        if (file.OwnerId != userId)
+        {
+            return Forbid();
         }
 
         await blobStorageService.DeleteAsync(file.BlobName,cancellationToken);
